Added SimplifiedDotsGameObserver and fixed the SimplifiedDotsGame.play method
SimplifiedDotsGameObserver is an observer to the SimplifiedDotsGame class,
which allows its implementations to keep track of who is added in the game,
who scored points and whenever the game is over (no more moves).

Maybe the SimplifiedDotsGame.play method should be reviewed, because of the
cells issue. We have a TODO for this in the code.